+++ core/domain/src/main/kotlin/com/nohjunh/android/watcha/assignment/core/domain/GetStorageTrackListUseCase.kt
@@ -1,14 +1,14 @@
 package com.nohjunh.android.watcha.assignment.core.domain
 
+import com.nohjunh.android.watcha.assignment.core.data.TrackRepository
 import kotlinx.coroutines.flow.Flow
-import com.nohjunh.android.watcha.assignment.core.data.StorageRepository
 import com.nohjunh.android.watcha.assignment.core.model.TrackItem
 import javax.inject.Inject
 
 class GetStorageTrackListUseCase @Inject constructor(
-    private val storageRepository: StorageRepository,
+    private val trackRepository: TrackRepository,
 ) {
     operator fun invoke(): Flow<List<TrackItem>> =
-        storageRepository.getStorageTrackList()
+        trackRepository.getStorageTrackList()
 
 }
