+++ core/domain/src/main/kotlin/com/nohjunh/android/watcha/assignment/core/domain/SaveTrackItemUseCase.kt
@@ -1,13 +1,13 @@
 package com.nohjunh.android.watcha.assignment.core.domain
 
+import com.nohjunh.android.watcha.assignment.core.data.TrackRepository
 import javax.inject.Inject
-import com.nohjunh.android.watcha.assignment.core.data.StorageRepository
 import com.nohjunh.android.watcha.assignment.core.model.TrackItem
 
 class SaveTrackItemUseCase @Inject constructor(
-    private val storageRepository: StorageRepository,
+    private val trackRepository: TrackRepository,
 ) {
     suspend operator fun invoke(trackItem: TrackItem): Result<Unit> =
-        storageRepository.saveTrackItem(trackItem = trackItem)
+        trackRepository.saveTrackItem(trackItem = trackItem)
 
 }
